1.1.4

+++ client/src/containers/FlightDirector/ClientsLobby.js
@@ -11,6 +11,8 @@ const FlightQuery = gql`
   query Flight {
     flights {
       id
+      name
+      flightType
       running
     }
   }
@@ -39,14 +41,23 @@ All information in all simulators in this flight will be reset.`
       });
     }
   };
-  const deleteFlight = () => {
+  const deleteFlight = flightType => () => {
     const flightId = props.match.params.flightId;
+
     if (
       window.confirm(
         `Are you sure you want to delete this flight?
 It will permenantly erase all simulators running in this flight.`
       )
     ) {
+      if (flightType) {
+        if (
+          !window.confirm(
+            `This flight hasn't been transmitted to SpaceEdventures.org. Without transmitting this flight, all participants will not receive credit for this flight. Are you sure you want to delete this flight before transmitting?`
+          )
+        )
+          return;
+      }
       const mutation = gql`
         mutation DeleteFlight($flightId: ID!) {
           deleteFlight(flightId: $flightId)
@@ -118,6 +129,14 @@ It will permenantly erase all simulators running in this flight.`
         )
       });
   };
+  const transmit = action => () => {
+    if (
+      window.confirm(`Are you sure you want to transmit this flight's information to SpaceEdVentures.org? 
+This can only be done once per flight and should only be done when the flight is complete.`)
+    ) {
+      action();
+    }
+  };
   return (
     <Query
       query={FlightQuery}
@@ -134,7 +153,7 @@ It will permenantly erase all simulators running in this flight.`
             <span>
               <div style={{ display: "flex", justifyContent: "space-between" }}>
                 <h4>
-                  Flight Lobby{" "}
+                  Flight Lobby: {flight.name}{" "}
                   <small>
                     <Link to="/">Return to Main</Link>
                   </small>
@@ -147,7 +166,7 @@ It will permenantly erase all simulators running in this flight.`
                 <Button
                   className="delete-flight"
                   color="danger"
-                  onClick={deleteFlight}
+                  onClick={deleteFlight(flight.flightType)}
                 >
                   Delete Flight
                 </Button>
@@ -214,11 +233,33 @@ It will permenantly erase all simulators running in this flight.`
                   }:${parseInt(window.location.port, 10) + 1}/exportFlight/${
                     props.match.params.flightId
                   }`}
-                  block
                   color="info"
                 >
                   Export Flight
                 </Button>
+                {flight.flightType && (
+                  <Mutation
+                    mutation={gql`
+                      mutation TransmitFlight($flightId: ID!) {
+                        assignSpaceEdventuresFlightRecord(flightId: $flightId)
+                      }
+                    `}
+                    variables={{ flightId: props.match.params.flightId }}
+                    refetchQueries={[
+                      {
+                        query: FlightQuery
+                      }
+                    ]}
+                  >
+                    {(action, { loading }) =>
+                      !loading && (
+                        <Button color="dark" onClick={transmit(action)}>
+                          Transmit to Space EdVentures
+                        </Button>
+                      )
+                    }
+                  </Mutation>
+                )}
               </ButtonGroup>
               <h5 className="text-right">
                 <Link
